+++ resources/views/themes/zeus/rain/widgets/FormsWidget.blade.php
@@ -1,0 +1,5 @@
+<div>
+    @if($data['form_slug'] !== null)
+        <livewire:bolt.fill-form :slug="$data['form_slug']" />
+    @endif
+</div>